requirements.txt
asgiref==3.6.0
Django==4.1.4
django-cleanup==6.0.0
django-filter==22.1
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
Pillow==9.4.0
pytz==2022.7
sqlparse==0.4.3
